+++ EpitechClient/Oz/Gyro.hh
@@ -14,15 +14,15 @@ class Gyro : public Unit
 public:
 	explicit Gyro(Gateway::Gateway & gateway);
 	void update() override;
-	int16_t getX() const noexcept;
-    int16_t getY() const noexcept;
-    int16_t getZ() const noexcept;
+	double getX() const noexcept;
+    double getY() const noexcept;
+    double getZ() const noexcept;
 
 private:
 	Gateway::Gateway & _gateway;
-    int16_t _xAngle;
-    int16_t _yAngle;
-    int16_t _zAngle;
+    double _xAngle;
+    double _yAngle;
+    double _zAngle;
 };
 
 }
